myPUPQC_Basic-main_v2-from-myPUPQCv2/requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.42.2
dj-database-url==2.3.0
Django==5.1.7
django-axes==7.0.2
django-cors-headers==4.7.0
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
frozenlist==1.5.0
geoip2==5.0.1
gunicorn==23.0.0
idna==3.10
maxminddb==2.6.3
multidict==6.1.0
openpyxl==3.1.5
packaging==24.2
pillow==11.1.0
propcache==0.3.0
psycopg2==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
user-agents==2.2.0
whitenoise==6.9.0
yarl==1.18.3
